feat: Added better support for moving collections between areas.
The areas-selector component now updated the category setting on the collection form.  The category is removed from the collection when the collection is removed from an area that is also the parent area of the category.  After this change, the UI now offers a select option in the collection form.