feat: Added new API controller function for removing collection from area prior to the colletion being assigned a category.
The default assumption is that the collection should belong to a collection group or category, but this may not always be the case. This addition allows the user to remove a collection from an area without first assigning a fake category.  The new function is private. The public API will always first attempt to remove the collection from the area and adjust the category relationship as needed.